api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.1
boto3==1.35.24
